requirements.txt
torch==1.13.1
torchvision==0.14.1
onnx==1.12.0
onnxruntime==1.13.1
fastapi==0.95.0
uvicorn==0.22.0
pillow==9.4.0
numpy==1.23.5
